pyproject.toml
[project]
name = "google-ads-reports"
version = "2.0.0"
description = "Simple Google Ads API client with automatic retries, exception handling, and flexible report generation"
authors = [{ name = "Joao Brito"}]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12,<3.13"
keywords = ["google-ads", "etl", "data-extraction", "reports", "serverless"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Typing :: Typed"
]
dependencies = [
    "google-ads>=24.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "mypy>=1.0.0",
    "types-PyYAML>=6.0.12.20240311",
    "types-protobuf>=4.24.0.20240417",
]

[project.urls]
Homepage = "https://github.com/machado000/google-ads-reports"
Documentation = "https://github.com/machado000/google-ads-reports#readme"
Issues = "https://github.com/machado000/google-ads-reports/issues"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "google_ads_reports" }]

[tool.mypy]
python_version = "3.12"
ignore_missing_imports = true
disallow_untyped_defs = true
check_untyped_defs = true
warn_unused_ignores = false
warn_return_any = true
show_error_codes = true
per_module_options = {"tests.*" = {disallow_untyped_defs = false}}